Client/requirements.txt
click==7.1.2
hidapi==0.12.0.post2
numpy==1.24.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt5-stubs==5.15.6.0
pyqt5-tools==5.15.4.3.2
python-dotenv==0.21.0
PyYAML==6.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
